TP_2/requirements.txt
#Python requirements lists.
pyfiglet
pandas
mysql.connector
getpass
requests
time
configparser
tqdm